se crea los permisos de productos y categoria

+++ app/Http/Controllers/ProductoController.php
@@ -15,6 +15,8 @@ class ProductoController extends Controller
      */
    public function index()
     {
+
+        $this->authorize('haveaccess','producto.index');
         $datos['productos']=Producto::paginate(5);
 
       return view('supervisor.producto.index',$datos);
@@ -27,6 +29,7 @@ class ProductoController extends Controller
      */
     public function create()
     {
+        $this->authorize('haveaccess','producto.create');
         return view('supervisor.producto.create');
     }
 
@@ -39,7 +42,7 @@ class ProductoController extends Controller
     public function store(Request $request)
     {
         //
-
+        $this->authorize('haveaccess','producto.create');
         // $datosCategorias=request()->all();
         $datosProductos=request()->except('_token');
         //validamos que el campo imagen tenga algo
@@ -61,6 +64,7 @@ class ProductoController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('haveaccess','producto.show');
         $producto= Producto::findOrFail($id);
 
         return view('supervisor.producto.show',compact('producto'));
@@ -74,6 +78,7 @@ class ProductoController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('haveaccess','producto.edit');
         //devuelve todo el valor del id.
         $producto= Producto::findOrFail($id);
 
@@ -89,7 +94,7 @@ class ProductoController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
+        $this->authorize('haveaccess','producto.edit');
         $datosProductos=request()->except(['_token','_method']);
 
        if ($request->hasFile('imagen')) {
@@ -116,6 +121,7 @@ class ProductoController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('haveaccess','producto.destroy');
         $producto= Producto::findOrFail($id);
 
        if(Storage::delete('public/'.$producto->imagen)){
